Math.Evaluation namespace is renamed to MathEvaluation to avoid conflicts with System.Math namespace
