+++ BenchmarkDotNet.Artifacts/results/Benchmarks-report.csv
@@ -1,7 +1,7 @@
 Method,Job,AnalyzeLaunchVariance,EvaluateOverhead,MaxAbsoluteError,MaxRelativeError,MinInvokeCount,MinIterationTime,OutlierMode,Affinity,EnvironmentVariables,Jit,LargeAddressAware,Platform,PowerPlanMode,Runtime,AllowVeryLargeObjects,Concurrent,CpuGroups,Force,HeapAffinitizeMask,HeapCount,NoAffinitize,RetainVm,Server,Arguments,BuildConfiguration,Clock,EngineFactory,NuGetReferences,IsMutator,InvocationCount,IterationCount,IterationTime,LaunchCount,MaxIterationCount,MaxWarmupIterationCount,MemoryRandomization,MinIterationCount,MinWarmupIterationCount,RunStrategy,UnrollFactor,WarmupCount,Mean,Error,StdDev,Median,Allocated
-"'MathEvaluator.Evaluate(""22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6"")'",.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,478.6224 ns,2.5991 ns,2.4312 ns,477.4653 ns,0 B
-"'Parse(""22888.32"") * Parse(""30"") / Parse(""323.34"") / Parse("".5"") - Parse(""-1"") / (Parse(""2"") + Parse(""22888.32"")) * Parse(""4"") - Parse(""6""))'",.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,288.1724 ns,4.2854 ns,3.7989 ns,287.7514 ns,0 B
-'22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6)',.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,0.0014 ns,0.0016 ns,0.0015 ns,0.0006 ns,0 B
-"'MathEvaluator.Evaluate(""22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6"")'",.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,446.6391 ns,1.6492 ns,1.5426 ns,446.7024 ns,0 B
-"'Parse(""22888.32"") * Parse(""30"") / Parse(""323.34"") / Parse("".5"") - Parse(""-1"") / (Parse(""2"") + Parse(""22888.32"")) * Parse(""4"") - Parse(""6""))'",.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,305.3880 ns,4.0496 ns,3.5899 ns,305.3351 ns,0 B
-'22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6)',.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,0.0046 ns,0.0067 ns,0.0060 ns,0.0019 ns,0 B
+"'MathEvaluator.Evaluate(""22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6"")'",.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,477.5644 ns,2.4669 ns,2.1869 ns,477.3509 ns,0 B
+"'Parse(""22888.32"") * Parse(""30"") / Parse(""323.34"") / Parse("".5"") - Parse(""-1"") / (Parse(""2"") + Parse(""22888.32"")) * Parse(""4"") - Parse(""6""))'",.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,288.4388 ns,2.0145 ns,1.8844 ns,288.3823 ns,0 B
+'22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6)',.NET 6.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 6.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,0.0000 ns,0.0000 ns,0.0000 ns,0.0000 ns,0 B
+"'MathEvaluator.Evaluate(""22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6"")'",.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,446.6624 ns,4.3340 ns,4.0540 ns,444.5351 ns,0 B
+"'Parse(""22888.32"") * Parse(""30"") / Parse(""323.34"") / Parse("".5"") - Parse(""-1"") / (Parse(""2"") + Parse(""22888.32"")) * Parse(""4"") - Parse(""6""))'",.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,304.8939 ns,2.7156 ns,2.5401 ns,304.6765 ns,0 B
+'22888.32 * 30 / 323.34 / .5 - -1 / (2 + 22888.32) * 4 - 6)',.NET 8.0,False,Default,Default,Default,Default,Default,Default,1111111111111111,Empty,RyuJit,Default,X64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,16,Default,0.0022 ns,0.0041 ns,0.0036 ns,0.0000 ns,0 B
