--- EvaluateMathExpression.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32228.430
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Math.Evaluation", "Math.Evaluation\Math.Evaluation.csproj", "{4C5614BF-1812-4B1E-82DB-1AFC8847EE0B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Math.Evaluation.Tests", "Math.Evaluation.Tests\Math.Evaluation.Tests.csproj", "{3B25C11F-B8F9-40AC-9105-9114B61132EF}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Math.Evaluation.Benchmarks", "Math.Evaluation.Benchmarks\Math.Evaluation.Benchmarks.csproj", "{3F1D2920-F9F7-4B50-A57E-A42F2A2110BF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4C5614BF-1812-4B1E-82DB-1AFC8847EE0B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C5614BF-1812-4B1E-82DB-1AFC8847EE0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C5614BF-1812-4B1E-82DB-1AFC8847EE0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C5614BF-1812-4B1E-82DB-1AFC8847EE0B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B25C11F-B8F9-40AC-9105-9114B61132EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B25C11F-B8F9-40AC-9105-9114B61132EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B25C11F-B8F9-40AC-9105-9114B61132EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B25C11F-B8F9-40AC-9105-9114B61132EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F1D2920-F9F7-4B50-A57E-A42F2A2110BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F1D2920-F9F7-4B50-A57E-A42F2A2110BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F1D2920-F9F7-4B50-A57E-A42F2A2110BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F1D2920-F9F7-4B50-A57E-A42F2A2110BF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8FB0293E-57DD-4400-B130-CCD9DC414CCB}
-	EndGlobalSection
-EndGlobal
