+++ MathEvaluation.Tests/MathEvaluation.Tests.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Math.Evaluation\Math.Evaluation.csproj" />
+    <ProjectReference Include="..\MathEvaluation\MathEvaluation.csproj" />
   </ItemGroup>
 
   <ItemGroup>
