+++ MathEvaluation.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32228.430
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathEvaluation", "MathEvaluation\MathEvaluation.csproj", "{9D7A102B-97EA-45DD-9F2D-26A4C6806BD2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathEvaluation.Benchmarks", "MathEvaluation.Benchmarks\MathEvaluation.Benchmarks.csproj", "{7579DBDC-4A6D-413B-91A5-0176F0E094EE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathEvaluation.Tests", "MathEvaluation.Tests\MathEvaluation.Tests.csproj", "{4418A2BC-8BD3-4D8C-A004-3E8ECA15AC78}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9D7A102B-97EA-45DD-9F2D-26A4C6806BD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D7A102B-97EA-45DD-9F2D-26A4C6806BD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D7A102B-97EA-45DD-9F2D-26A4C6806BD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D7A102B-97EA-45DD-9F2D-26A4C6806BD2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7579DBDC-4A6D-413B-91A5-0176F0E094EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7579DBDC-4A6D-413B-91A5-0176F0E094EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7579DBDC-4A6D-413B-91A5-0176F0E094EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7579DBDC-4A6D-413B-91A5-0176F0E094EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4418A2BC-8BD3-4D8C-A004-3E8ECA15AC78}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4418A2BC-8BD3-4D8C-A004-3E8ECA15AC78}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4418A2BC-8BD3-4D8C-A004-3E8ECA15AC78}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4418A2BC-8BD3-4D8C-A004-3E8ECA15AC78}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8FB0293E-57DD-4400-B130-CCD9DC414CCB}
+	EndGlobalSection
+EndGlobal
