+++ MathEvaluation/MathEvaluation.csproj
